RegularPolygons functioning

+++ ui-shapes/src/main/java/com/aleksejb/shapes/ShapesEvent.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.graphics.Color
 import com.aleksejb.domain_core.model.OutlineType
 
 sealed interface ShapesEvent {
-    data class OnNumberOfSidesSelected(val numberOfSides: Int): ShapesEvent
+    data class OnNumberOfSidesSelected(val numberOfSides: Float): ShapesEvent
     data class OnOutlineTypeSelected(val outlineType: OutlineType): ShapesEvent
     data class OnFillSelected(val fillColor: Color): ShapesEvent
 }
